V.0.5.4.0.4: полное сохранение проекта; версии 0.5.4.0.4 (version.txt, pyproject, constants, README); mypy-доводка: 24 ошибки в 14 файлах закрыты без вмешательства в логику

+++ backend/app/services/calibration/metrics/character_change.py
@@ -59,7 +59,7 @@ class CharacterChangeRate(CalibrationMetric):
     def compute(self) -> Optional[float]:
         if not self._deltas:
             return None
-        return sum(self._deltas) / len(self._deltas)
+        return float(sum(self._deltas) / len(self._deltas))
 
     def reset(self) -> None:
         self._prev.clear()
